Discontinued authorization added

+++ src/main/java/com/spartaglobal/musicapiproject/controllers/DiscontinuedTrackController.java
@@ -1,12 +1,11 @@
 package com.spartaglobal.musicapiproject.controllers;
-import com.spartaglobal.musicapiproject.entities.Album;
 import com.spartaglobal.musicapiproject.entities.DiscontinuedTrack;
-import com.spartaglobal.musicapiproject.entities.Track;
 import com.spartaglobal.musicapiproject.repositories.DiscontinuedTrackRepository;
+import com.spartaglobal.musicapiproject.services.AuthorizationService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
-import javax.validation.Valid;
 import java.util.List;
 import java.util.Optional;
 
@@ -16,6 +15,9 @@ public class DiscontinuedTrackController {
     @Autowired
     private DiscontinuedTrackRepository discontinuedTrackRepository;
 
+    @Autowired
+    private AuthorizationService authorizationService;
+
     @GetMapping(value = "/chinook/discontinuedtrack")
     public Optional<DiscontinuedTrack> getDiscontinuedTrack(@RequestParam Integer id) {
         Optional<DiscontinuedTrack> result = discontinuedTrackRepository.findById(id);
@@ -23,13 +25,23 @@ public class DiscontinuedTrackController {
     }
 
     @DeleteMapping(value = "/chinook/discontinuedtrack/delete")
-    public void deleteDiscontinuedTrack(@RequestParam Integer id) {
+    public ResponseEntity deleteDiscontinuedTrack(@RequestParam Integer id, @RequestHeader("Authorization") String authTokenHeader) {
+        String token = authTokenHeader.split(" ")[1];
+        if (authorizationService.isAuthorizedForAction(token, "chinook/discontinuedtrack/delete")) {
+            return new ResponseEntity<>("Not Authorized", HttpStatus.UNAUTHORIZED);
+        }
         discontinuedTrackRepository.deleteById(id);
+        return new ResponseEntity("Track removed from discontinued", HttpStatus.OK);
     }
 
     @PostMapping(value = "/chinook/discontinuedtrack/insert")
-    public DiscontinuedTrack insertDiscontinuedTrack(@RequestBody DiscontinuedTrack track) {
-        return discontinuedTrackRepository.save(track);
+    public ResponseEntity insertDiscontinuedTrack(@RequestBody DiscontinuedTrack track, @RequestHeader("Authorization") String authTokenHeader) {
+        String token = authTokenHeader.split(" ")[1];
+        if (authorizationService.isAuthorizedForAction(token, "chinook/discontinuedtrack/insert")) {
+            return new ResponseEntity<>("Not Authorized", HttpStatus.UNAUTHORIZED);
+        }
+        discontinuedTrackRepository.save(track);
+        return new ResponseEntity("Track inserted into discontinued", HttpStatus.OK);
     }
 
     @GetMapping(value = "/chinook/discontinuedtracks")
@@ -37,5 +49,3 @@ public class DiscontinuedTrackController {
         return discontinuedTrackRepository.findAll();
     }
 }
-
-
